Major Refactoring focused on CloudInit vs. remote execution.
This removes the airflow module which was entirely for remote-exec based deployment.  Instead the deployment uses an init script on the bastion host.

This bastion host can now be put into a private subnet (no internet exposure) and deployment will still work.

In addition, set the OKE deployment to default to VCN based (private) endpoint, with an opt-in for public endpoint.

Also tweaked a few Docker container build options.